requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
NaverTTS==0.3.1
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
urllib3==2.2.2
yarl==1.9.4
